+++ src/main/java/com/supermartijn642/simplemagnets/SimpleMagnets.java
@@ -117,6 +117,7 @@ public class SimpleMagnets implements ModInitializer {
     private static void registerGenerators(){
         GeneratorRegistrationHandler handler = GeneratorRegistrationHandler.get("simplemagnets");
         handler.addGenerator(SimpleMagnetsModelGenerator::new);
+        handler.addGenerator(SimpleMagnetsAtlasSourceGenerator::new);
         handler.addGenerator(SimpleMagnetsBlockStateGenerator::new);
         handler.addGenerator(SimpleMagnetsItemInfoGenerator::new);
         handler.addGenerator(SimpleMagnetsLanguageGenerator::new);